Code index.html and add content for all sections
Created the structure and layout using code for all the sections
expect for form and footer. Needed more knowledge and thought
on how to create forms and footers using html.

In the sections expect for the above mentioned, the layouts have
been created and all the content has been added.